Add directory user events
