+++ workos/resources/events.py
@@ -3,12 +3,16 @@ from typing_extensions import Annotated
 from pydantic import Field
 from workos.resources.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_group import DirectoryGroup
+from workos.types.directory_sync.directory_user import DirectoryUser
 from workos.types.events.directory_group_with_previous_attributes import (
     DirectoryGroupWithPreviousAttributes,
 )
 from workos.types.events.directory_payload import DirectoryPayload
 from workos.types.events.directory_payload_with_legacy_fields import (
     DirectoryPayloadWithLegacyFields,
+)
+from workos.types.events.directory_user_with_previous_attributes import (
+    DirectoryUserWithPreviousAttributes,
 )
 from workos.typing.literals import LiteralOrUntyped
 
@@ -18,6 +22,9 @@ EventType = Literal[
     "dsync.group.created",
     "dsync.group.deleted",
     "dsync.group.updated",
+    "dsync.user.created",
+    "dsync.user.deleted",
+    "dsync.user.updated",
 ]
 EventTypeDiscriminator = TypeVar("EventTypeDiscriminator", bound=EventType)
 EventPayload = TypeVar(
@@ -26,6 +33,7 @@ EventPayload = TypeVar(
     DirectoryPayloadWithLegacyFields,
     DirectoryGroup,
     DirectoryGroupWithPreviousAttributes,
+    DirectoryUser,
 )
 
 
@@ -68,7 +76,25 @@ class DirectoryGroupDeletedEvent(
 class DirectoryGroupUpdatedEvent(
     EventModel[Literal["dsync.group.updated"], DirectoryGroupWithPreviousAttributes]
 ):
-    event: Literal["dsync.group.deleted"]
+    event: Literal["dsync.group.updated"]
+
+
+class DirectoryUserCreatedEvent(
+    EventModel[Literal["dsync.user.created"], DirectoryUser]
+):
+    event: Literal["dsync.user.created"]
+
+
+class DirectoryUserDeletedEvent(
+    EventModel[Literal["dsync.user.deleted"], DirectoryUser]
+):
+    event: Literal["dsync.user.deleted"]
+
+
+class DirectoryUserUpdatedEvent(
+    EventModel[Literal["dsync.user.updated"], DirectoryUserWithPreviousAttributes]
+):
+    event: Literal["dsync.user.updated"]
 
 
 Event = Annotated[
