+++ workos/types/events/directory_group_with_previous_attributes.py
@@ -1,4 +1,3 @@
-from workos.resources.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_group import DirectoryGroup
 from workos.types.events.previous_attributes import PreviousAttributes
