+++ workos/types/events/directory_user_with_previous_attributes.py
@@ -1,0 +1,6 @@
+from workos.types.directory_sync.directory_user import DirectoryUser
+from workos.types.events.previous_attributes import PreviousAttributes
+
+
+class DirectoryUserWithPreviousAttributes(DirectoryUser):
+    previous_attributes: PreviousAttributes
